should be fine, needs last revision before sumbitting

+++ ft_strncmp.c
@@ -6,15 +6,15 @@
 /*   By: kael-mal <kael-mal@student.1337.ma>        +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2025/10/20 11:00:28 by kael-mal          #+#    #+#             */
-/*   Updated: 2025/10/20 11:00:30 by kael-mal         ###   ########.fr       */
+/*   Updated: 2025/10/20 18:17:07 by kael-mal         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "libft.h"
 
-int	ft_strncmp(const char * s1, const char *s2, size_t n)
+int	ft_strncmp(const char *s1, const char *s2, size_t n)
 {
-	int	i;
+	size_t	i;
 
 	i = 0;
 	while (s1[i] == s2[i] && s1[i] && s2[i] && i < n -1)
